Moved LocationResource and EC2InstanceResource classes to their own apps.
